moved library files into shlib
