+++ shlib/cfg_format.py
