+++ shlib/cfg_formatter.py
