+++ shlib/fmt_config.py
