+++ shlib/include_loader.py
