+++ test/test_files.py
@@ -2,9 +2,9 @@ import sys
 import os
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
-import include_loader
-import fmt_config
-import cfg_format
-import logger
+from shlib import include_loader
+from shlib import fmt_config
+from shlib import cfg_format
+from shlib import logger
 
 logger.set_test_mode()
