+++ test/test_include_loader.py
@@ -2,7 +2,7 @@ import sys
 import os
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
-from include_loader import IncludeLoader, YmlImportException
+from shlib.include_loader import IncludeLoader, YmlImportException
 import unittest
 import yaml
